requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
deepface==0.0.49
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
gdown==3.12.2
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mtcnn==0.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.3
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
